requirements.txt
streamlit
feedparser
newspaper3k
lxml_html_clean
lxml